fix "Couldn't parse ISO 8601 date string" error
"Jellyfin 10.10.0 requires the date in the manifest.json for plugins to be in ISO 8601 format instead of a standard date and time format. A fix has already been implemented to support plugin revisions with incorrect timestamps. This will be included in the first bug fix release (10.10.1), though a specific timeline for that release has not yet been announced."

https://forum.jellyfin.org/t-merge-versions-plugin-jellyfin-10-10-0-error?pid=40690#pid40690